Move check_qualification to Stadium
